+++ src/Symfony/Component/Notifier/Bridge/Sendberry/Tests/SendberryTransportTest.php
@@ -16,28 +16,30 @@ use Symfony\Component\Notifier\Message\ChatMessage;
 use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 final class SendberryTransportTest extends TransportTestCase
 {
-    public function createTransport(HttpClientInterface $client = null): SendberryTransport
+    public static function createTransport(HttpClientInterface $client = null): SendberryTransport
     {
-        return new SendberryTransport('username', 'password', 'auth_key', 'from', $client ?? $this->createMock(HttpClientInterface::class));
+        return new SendberryTransport('username', 'password', 'auth_key', 'from', $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['sendberry://username:password@api.sendberry.com?auth_key=auth_key&from=from', $this->createTransport()];
+        yield ['sendberry://username:password@api.sendberry.com?auth_key=auth_key&from=from', self::createTransport()];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('+0611223344', 'Hello!')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 }