+++ src/Symfony/Component/Notifier/Bridge/Slack/Tests/SlackTransportTest.php
@@ -23,31 +23,34 @@ use Symfony\Component\Notifier\Message\MessageOptionsInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Notification\Notification;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
+use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 final class SlackTransportTest extends TransportTestCase
 {
-    public function createTransport(HttpClientInterface $client = null, string $channel = null): SlackTransport
+    public static function createTransport(HttpClientInterface $client = null, string $channel = null): SlackTransport
     {
-        return new SlackTransport('xoxb-TestToken', $channel, $client ?? $this->createMock(HttpClientInterface::class));
+        return new SlackTransport('xoxb-TestToken', $channel, $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['slack://slack.com', $this->createTransport()];
-        yield ['slack://slack.com?channel=test+Channel', $this->createTransport(null, 'test Channel')];
+        yield ['slack://slack.com', self::createTransport()];
+        yield ['slack://slack.com?channel=test+Channel', self::createTransport(null, 'test Channel')];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('0611223344', 'Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 
     public function testInstatiatingWithAnInvalidSlackTokenThrowsInvalidArgumentException()
@@ -74,7 +77,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, 'testChannel');
+        $transport = self::createTransport($client, 'testChannel');
 
         $transport->send(new ChatMessage('testMessage'));
     }
@@ -97,7 +100,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, 'testChannel');
+        $transport = self::createTransport($client, 'testChannel');
 
         $transport->send(new ChatMessage('testMessage'));
     }
@@ -125,7 +128,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, $channel);
+        $transport = self::createTransport($client, $channel);
 
         $sentMessage = $transport->send(new ChatMessage('testMessage'));
 
@@ -163,7 +166,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, $channel);
+        $transport = self::createTransport($client, $channel);
 
         $sentMessage = $transport->send($chatMessage);
 
@@ -178,7 +181,7 @@ final class SlackTransportTest extends TransportTestCase
             return $this->createMock(ResponseInterface::class);
         });
 
-        $transport = $this->createTransport($client, 'testChannel');
+        $transport = self::createTransport($client, 'testChannel');
 
         $transport->send(new ChatMessage('testMessage', $this->createMock(MessageOptionsInterface::class)));
     }
@@ -208,7 +211,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, $channel);
+        $transport = self::createTransport($client, $channel);
 
         $transport->send(new ChatMessage('testMessage'));
     }
@@ -231,7 +234,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client);
+        $transport = self::createTransport($client);
 
         $transport->send(new ChatMessage('testMessage'));
     }
@@ -256,7 +259,7 @@ final class SlackTransportTest extends TransportTestCase
             return $response;
         });
 
-        $transport = $this->createTransport($client, 'testChannel');
+        $transport = self::createTransport($client, 'testChannel');
 
         $this->expectException(TransportException::class);
         $this->expectExceptionMessage('Unable to post the Slack message: "invalid_blocks" (no more than 50 items allowed [json-pointer:/blocks]).');