+++ src/Symfony/Component/Notifier/Bridge/Zendesk/Tests/ZendeskTransportTest.php
@@ -17,29 +17,31 @@ use Symfony\Component\Notifier\Message\ChatMessage;
 use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
+use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 final class ZendeskTransportTest extends TransportTestCase
 {
-    public function createTransport(HttpClientInterface $client = null): ZendeskTransport
+    public static function createTransport(HttpClientInterface $client = null): ZendeskTransport
     {
-        return (new ZendeskTransport('testEmail', 'testToken', $client ?? $this->createMock(HttpClientInterface::class)))->setHost('test.zendesk.com');
+        return (new ZendeskTransport('testEmail', 'testToken', $client ?? new DummyHttpClient()))->setHost('test.zendesk.com');
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['zendesk://test.zendesk.com', $this->createTransport()];
+        yield ['zendesk://test.zendesk.com', self::createTransport()];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
         yield [new ChatMessage('Hello!', new ZendeskOptions('urgent'))];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('0611223344', 'Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
     }
 }