+++ src/Symfony/Component/Notifier/CHANGELOG.md
@@ -12,6 +12,12 @@ CHANGELOG
 
  * Use importance level to set flash message type
 
+5.4.21
+------
+
+ * [BC BREAK] The following data providers for `TransportTestCase` are now static: `toStringProvider()`, `supportedMessagesProvider()` and `unsupportedMessagesProvider()`
+ * [BC BREAK] `TransportTestCase::createTransport()` is now static
+
 5.4
 ---